+++ docs/src/markdown/about/changelog.md
@@ -2,6 +2,7 @@
 
 ## 2.4
 
+- **NEW**: Add Rec. 709 RGB color space.
 - **NEW**: Add the 1960 UCS color space.
 - **NEW**: Add correlated color temperature support with new `cct()` and `blackbody()` API.
 - **NEW**: Add support Robertson 1968 and Ohno 2013 CCT plugins.